src/requirements.txt
# src/requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
requests>=2.31.0
google-generativeai>=0.7.2
pypdf>=4.2.0
python-dotenv>=1.0.1
